Make all reachability-status callbacks from the main GCD queue.
Because the run-loop reachability callback executed its callback
directly, whereas the initial check (fired off from
`-[AFNetworkReachabilityManager startMonitoring]`) executed the callback
from a GCD queue, it was possible for the callbacks to be executed out
of order from when the status was obtained, leaving the listener in the
incorrect state.

By sending both routes through the same GCD-queuing function, both the
callbacks and NSNotifications will be processed in the order in which
they are sent. It also removes some repetition in the code.